# SI/si-ii-tp1-groupe5-dotnet-22-23/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SI-II TP1 Groupe 5 22-23",
    "description": "API Specification for SI II TP1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1/"
    }
  ],
...
  "paths": {
...
    "/matches": {
      "get": {
        "tags": [
          "Match"
        ],
        "summary": "Get all matches",
        "operationId": "getMatches",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Match"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "post": {
        "tags": [
          "Match"
        ],
        "summary": "Add a new match to the store",
        "operationId": "addMatch",
        "requestBody": {
          "description": "Match object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Match"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/matches/{matchId}": {
      "get": {
        "tags": [
          "Match"
        ],
        "summary": "Find match by ID",
        "description": "Returns a single match",
        "operationId": "getMatchById",
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "description": "ID of match to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Match"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Match not found"
          }
        }
      },
      "put": {
        "tags": [
          "Match"
        ],
        "summary": "Update an existing match",
        "operationId": "updateMatch",
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "description": "ID of match to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Match object that needs to be added to the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Match"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Match not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
  },
...
}